#530: Move tests to correct folder
