+++ Api/Xipona.Api.Endpoints.Tests/v1/Endpoints/ItemCategoryTests/GetItemCategoryByIdAsyncTests.cs
@@ -10,7 +10,7 @@ using ProjectHermes.Xipona.Api.Endpoint.v1.Endpoints;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemCategoryEndpointTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Endpoints.ItemCategoryEndpointTests;
 
 public class GetItemCategoryByIdTests : EndpointQueryNoConverterTestsBase<
     ItemCategoryByIdQuery, IItemCategory, ItemCategoryContract,
