+++ Api/Xipona.Api.Endpoints.Tests/v1/Endpoints/ItemCategoryTests/ModifyItemCategoryTests.cs
@@ -12,7 +12,7 @@ using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
 using System.Net.Http;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemCategoryEndpointTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Endpoints.ItemCategoryEndpointTests;
 
 public class ModifyItemCategoryTests : EndpointCommandTestsBase<ModifyItemCategoryContract,
     ModifyItemCategoryCommand, bool, ModifyItemCategoryTests.ModifyItemCategoryFixture>
