+++ Api/Xipona.Api.Endpoints.Tests/v1/Endpoints/ItemTests/GetItemTypePricesTests.cs
@@ -1,33 +1,28 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
-using ProjectHermes.Xipona.Api.ApplicationServices.Items.Queries.SearchItemsByItemCategory;
+using ProjectHermes.Xipona.Api.ApplicationServices.Items.Queries.GetItemTypePrices;
 using ProjectHermes.Xipona.Api.Contracts.Common;
-using ProjectHermes.Xipona.Api.Contracts.Items.Queries.SearchItemsByItemCategory;
+using ProjectHermes.Xipona.Api.Contracts.Items.Queries.GetItemTypePrices;
 using ProjectHermes.Xipona.Api.Domain.Common.Reasons;
-using ProjectHermes.Xipona.Api.Domain.ItemCategories.Models;
-using ProjectHermes.Xipona.Api.Domain.Items.Services.Searches;
+using ProjectHermes.Xipona.Api.Domain.Items.Models;
+using ProjectHermes.Xipona.Api.Domain.Items.Services.Queries;
+using ProjectHermes.Xipona.Api.Domain.Stores.Models;
 using ProjectHermes.Xipona.Api.Endpoint.v1.Endpoints;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemControllerTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Endpoints.ItemTests;
 
-public class SearchItemsByItemCategoryTests : EndpointEnumerableQueryNoConverterTestsBase<
-    SearchItemsByItemCategoryQuery, SearchItemByItemCategoryResult, SearchItemByItemCategoryResultContract,
-    SearchItemsByItemCategoryTests.SearchItemsByItemCategoryFixture>
+public class GetItemTypePricesTests() :
+    EndpointQueryNoConverterTestsBase<GetItemTypePricesQuery, ItemTypePricesReadModel, ItemTypePricesContract,
+        GetItemTypePricesTests.GetItemTypePricesFixture>(new GetItemTypePricesFixture())
 {
-    public SearchItemsByItemCategoryTests() : base(new SearchItemsByItemCategoryFixture())
-    {
-    }
-
     [Theory]
-    [InlineData(ErrorReasonCode.StoresNotFound)]
-    [InlineData(ErrorReasonCode.ItemCategoryNotFound)]
+    [InlineData(ErrorReasonCode.ItemNotFound)]
     public async Task EndpointCall_WithDomainException_ShouldReturnNotFound(ErrorReasonCode errorCode)
     {
         // Arrange
-        Fixture.SetupParameters();
         Fixture.SetupQuery();
         Fixture.SetupDomainException(errorCode);
         Fixture.SetupDomainExceptionInQueryDispatcher();
@@ -43,24 +38,23 @@ public class SearchItemsByItemCategoryTests : EndpointEnumerableQueryNoConverter
         unprocessableEntity!.Value.Should().BeEquivalentTo(Fixture.ExpectedErrorContract);
     }
 
-    public sealed class SearchItemsByItemCategoryFixture : EndpointEnumerableQueryNoConverterFixtureBase
+    public sealed class GetItemTypePricesFixture : EndpointQueryNoConverterFixtureBase
     {
-        private readonly ItemCategoryId _itemCategoryId = ItemCategoryId.New;
+        private readonly Guid _itemId = Guid.NewGuid();
+        private readonly Guid _storeId = Guid.NewGuid();
 
-        public SearchItemsByItemCategoryFixture()
+        public GetItemTypePricesFixture()
         {
             PossibleResultsList.Add(new OkStatusResult());
-            PossibleResultsList.Add(new NoContentStatusResult());
-            PossibleResultsList.Add(new UnprocessableEntityStatusResult(
-                ErrorReasonCode.ItemCategoryNotFound, ErrorReasonCode.StoresNotFound));
             PossibleResultsList.Add(new NotFoundStatusResult());
+            PossibleResultsList.Add(new UnprocessableEntityStatusResult(ErrorReasonCode.ItemNotFound));
         }
 
-        public override string RoutePattern => "/v1/items/search/by-item-category/{itemCategoryId:guid}";
+        public override string RoutePattern => "/v1/items/{id:guid}/type-prices";
 
-        public override async Task<IResult> ExecuteTestMethod()
+        public override Task<IResult> ExecuteTestMethod()
         {
-            return await ItemEndpoints.SearchItemsByItemCategory(_itemCategoryId,
+            return ItemEndpoints.GetItemTypePrices(_itemId, _storeId,
                 QueryDispatcherMock.Object,
                 ContractConverterMock.Object,
                 ErrorConverterMock.Object,
@@ -78,7 +72,7 @@ public class SearchItemsByItemCategoryTests : EndpointEnumerableQueryNoConverter
 
         public override void SetupQuery()
         {
-            Query = new SearchItemsByItemCategoryQuery(_itemCategoryId);
+            Query = new GetItemTypePricesQuery(new ItemId(_itemId), new StoreId(_storeId));
         }
     }
 }
