+++ Api/Xipona.Api.Endpoints.Tests/v1/Endpoints/ItemTests/SearchItemsTests.cs
@@ -9,7 +9,7 @@ using ProjectHermes.Xipona.Api.Endpoints.Tests.Common;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemControllerTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Endpoints.ItemTests;
 
 public class SearchItemsTests :
     EndpointEnumerableQueryNoConverterTestsBase<SearchItemQuery, SearchItemResultReadModel, SearchItemResultContract,
