+++ Api/Xipona.Api.Endpoints.Tests/v1/Endpoints/ItemTests/UpdateItemPriceTests.cs
@@ -13,7 +13,7 @@ using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
 using System.Net.Http;
 
-namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Controllers.ItemControllerTests;
+namespace ProjectHermes.Xipona.Api.Endpoints.Tests.v1.Endpoints.ItemTests;
 
 public class UpdateItemPriceTests : EndpointCommandTestsBase<(Guid, UpdateItemPriceContract), UpdateItemPriceCommand, bool,
     UpdateItemPriceTests.UpdateItemPriceFixture>
